Execute the Study

Execute the Study

execute(
  connectionDetails,
  cdmDatabaseSchema,
  cdmDatabaseName = "friendly database name",
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  oracleTempSchema = cohortDatabaseSchema,
  outputFolder,
  createProtocol = F,
  createCohorts = F,
  runDiagnostic = F,
  viewDiagnostic = F,
  runAnalyses = F,
  createResultsDoc = F,
  createValidationPackage = F,
  analysesToValidate = NULL,
  packageResults = F,
  minCellCount = 5,
  createShiny = F,
  createJournalDocument = F,
  analysisIdDocument = 1,
  verbosity = "INFO",
  cdmVersion = 5,
  cohortVariableSetting = NULL
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package.

cdmDatabaseSchema
Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

cdmDatabaseName
Shareable name of the database

cohortDatabaseSchema
Schema name where intermediate data can be stored. You will need to have write priviliges in this schema. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

cohortTable
The name of the table that will be created in the work database schema. This table will hold the target population cohorts used in this study.

oracleTempSchema
Should be used in Oracle to specify a schema where the user has write priviliges for storing temporary tables.

outputFolder
Name of local folder to place results; make sure to use forward slashes (/). Do not use a folder on a network drive since this greatly impacts performance.

createProtocol
Creates a protocol based on the analyses specification

createCohorts
Create the cohortTable table with the target population and outcome cohorts?

runDiagnostic
Runs a diagnostic of the T, O and tar settings for the cdmDatabaseSchema - can be used to check whether to change settings or whether the prediction may not do well.

viewDiagnostic
Opens a shiny app with the diagnostic results (run after runDiagnostic completes)

runAnalyses
Run the model development

createResultsDoc
Create a document containing the results of each prediction

createValidationPackage
Create a package for sharing the models

analysesToValidate
A vector of analysis ids (e.g., c(1,3,10)) specifying which analysese to export into validation package. Default is NULL and all are exported.

packageResults
Should results be packaged for later sharing?

minCellCount
The minimum number of subjects contributing to a count before it can be included in packaged results.

createShiny
Create a shiny app with the results

createJournalDocument
Do you want to create a template journal document populated with results?

analysisIdDocument
Which Analysis_id do you want to create the document for?

verbosity
Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:

DEBUGHighest verbosity showing all debug statements

TRACEShowing information about start and end of steps

INFOShow informative information (Default)

WARNShow warning messages

ERRORShow error messages

FATALBe silent except for fatal errors

cdmVersion
The version of the common data model

cohortVariableSetting
the name of the custom cohort covariate settings to use

Details

This function executes the SkeletonPredictionStudy Study.

Examples